requirements.txt
fastapi
uvicorn
pydantic
ddgs
llama-cpp-python